+++ Src/Expression.cs
@@ -4,24 +4,33 @@
 
 using LiteDB;
 using System;
+using System.Linq;
 using System.Collections;
 
 namespace Ldbc
 {
 	public sealed class Expression
 	{
-		public BsonExpression BsonExpression { get; private set; }
-
+		readonly BsonExpression _expression;
 		Dictionary _Parameters;
+
 		public Dictionary Parameters
 		{
 			get
 			{
-				if (_Parameters == null)
-					_Parameters = new Dictionary(BsonExpression.Parameters);
-
-				return _Parameters;
+				return _Parameters ?? (_Parameters = new Dictionary(_expression.Parameters));
 			}
+		}
+
+		[Obsolete("Designed for scripts.")]
+		public static implicit operator BsonExpression(Expression value)
+		{
+			return value?._expression;
+		}
+
+		public BsonExpression ToBsonExpression()
+		{
+			return _expression;
 		}
 
 		public Expression(string expression)
@@ -29,12 +38,12 @@ namespace Ldbc
 			if (expression == null)
 				throw new ArgumentNullException(nameof(expression));
 
-			BsonExpression = BsonExpression.Create(expression);
+			_expression = BsonExpression.Create(expression);
 		}
 
 		public Expression(BsonExpression expression)
 		{
-			BsonExpression = expression ?? throw new ArgumentNullException(nameof(expression));
+			_expression = expression ?? throw new ArgumentNullException(nameof(expression));
 		}
 
 		public Expression(string expression, IDictionary parameters)
@@ -45,25 +54,76 @@ namespace Ldbc
 				throw new ArgumentNullException(nameof(parameters));
 
 			var parameters2 = Actor.ToBsonDocumentFromDictionary(parameters);
-			BsonExpression = BsonExpression.Create(expression, parameters2);
+			_expression = BsonExpression.Create(expression, parameters2);
 		}
 
-		internal static Expression Create(object value)
+		public Expression(string expression, params object [] args)
+		{
+			if (expression == null)
+				throw new ArgumentNullException(nameof(expression));
+			if (args == null)
+				throw new ArgumentNullException(nameof(args));
+
+			var args2 = new BsonValue[args.Length];
+			for(var i = 0; i < args.Length; ++i)
+				args2[i] = Actor.ToBsonValue(args[i]);
+
+			_expression = BsonExpression.Create(expression, args2);
+		}
+
+		public override string ToString()
+		{
+			return _expression.ToString();
+		}
+
+		static void SetParameters(BsonDocument that, IDictionary parameters)
+		{
+			if (parameters == null)
+				throw new ArgumentNullException(nameof(parameters));
+
+			foreach (DictionaryEntry kv in parameters)
+				that[kv.Key.ToString()] = Actor.ToBsonValue(kv.Value);
+		}
+
+		static void SetArguments(BsonDocument that, params object[] args)
+		{
+			if (args == null)
+				throw new ArgumentNullException(nameof(args));
+
+			for (var i = 0; i < args.Length; ++i)
+				that[$"{i}"] = Actor.ToBsonValue(args[i]);
+		}
+
+		internal static BsonExpression Input(object value)
 		{
 			if (value == null)
 				throw new ArgumentNullException(nameof(value));
 
-			if (value is Expression expr1)
+			if (value is string text)
+				return BsonExpression.Create(text);
+
+			if (value is BsonExpression expr1)
 				return expr1;
 
-			if (value is BsonExpression expr2)
-				return new Expression(expr2);
+			if (value is Expression expr2)
+				return expr2.ToBsonExpression();
 
-			if (value is string text)
-				return new Expression(text);
+			if (value is IList list)
+			{
+				if (list.Count == 2 && list[1] is IDictionary dic)
+				{
+					var expr = Input(list[0]);
+					SetParameters(expr.Parameters, dic);
+					return expr;
+				}
 
-			if (value is IList list && list.Count == 2)
-				return new Expression((string)list[0], (IDictionary)list[1]);
+				if (list.Count >= 2)
+				{
+					var expr = Input(list[0]);
+					SetArguments(expr.Parameters, list.Cast<object>().Skip(1).ToArray());
+					return expr;
+				}
+			}
 
 			throw new ArgumentException(Res.CannotConvert2(value, "expression"));
 		}