+++ Tests/Class.test.ps1
@@ -54,28 +54,6 @@ task ClassExample {
 	equals $r[0].Location.Line 11
 }
 
-#! https://github.com/mbdavid/LiteDB/issues/1483
-#! auto id is currently not working as expected
-#! when resolved, remove my work around
-task TODO {
-	Use-LiteDatabase :memory: {
-		$test = Get-LiteCollection Test Int32
-		try {
-			# InsertBulk is fine
-			Add-LiteData $test @{_id = 0} -Bulk
-
-			# Insert has issue
-			$d = [LiteDB.BsonDocument]::new()
-			$d["_id"] = 0
-			$test.Insert($d)
-			throw
-		}
-		catch {
-			assert ("$_" -like '*Object reference not set to an instance of an object.*')
-		}
-	}
-}
-
 task DbRef {
 	$customer = [Customer]@{CustomerId = 1; Name = 'John'}
 	$product1 = [Product]@{ProductId = 1; Name = 'Foo'}