+++ Release-Notes.md
@@ -1,5 +1,21 @@
 # Ldbc Release Notes
 
+## v0.6.3
+
+LiteDB 5.0.3
+
+`Test-LiteData` - new parameter `ById`.
+
+Parametrized expressions - support both parameters and arguments notations:
+
+```powershell
+# expression followed by IDictionary with named parameters
+X-LiteData .. -Where ('$x = @x AND $y = @y', @{x = 1; y = 2})
+
+# expression followed by indexed positional arguments
+X-LiteData .. -Where ('$x = @0 AND $y = @1', 1, 2)
+```
+
 ## v0.6.2
 
 `Get-LiteData`